frontend-app-api: deduplicate joinPaths utility in routing
Extract the identical joinPaths function from both
extractRouteInfoFromAppNode.ts and RouteResolver.ts into a shared
joinPaths.ts module, then import it from both consumers.

Made-with: Cursor
